requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
python-multipart>=0.0.5
thefuzz>=0.20.0
python-Levenshtein>=0.23.0
